+++ app/src/main/java/ch/epfl/sweng/eventmanager/repository/ScheduledItemRepository.java
@@ -15,35 +15,15 @@ import java.util.List;
  * @author Louis Vialar
  */
 @Singleton
-public class ScheduledItemRepository {
+public class ScheduledItemRepository extends AbstractEventDataRepository<ScheduledItem> {
     private static final String TAG = "ScheduledItemRepository";
 
     @Inject
     public ScheduledItemRepository() {
+        super("schedule_items", new GenericTypeIndicator<List<ScheduledItem>>() {});
     }
 
     public LiveData<List<ScheduledItem>> getConcerts(int eventId) {
-        final MutableLiveData<List<ScheduledItem>> data = new MutableLiveData<>();
-        DatabaseReference dbRef = FirebaseDatabase.getInstance()
-                .getReference("schedule_items")
-                .child("event_" + eventId);
-
-        dbRef.addValueEventListener(new ValueEventListener() {
-            @Override
-            public void onDataChange(@NonNull DataSnapshot dataSnapshot) {
-                GenericTypeIndicator<List<ScheduledItem>> typeToken = new GenericTypeIndicator<List<ScheduledItem>>() {
-                };
-
-                List<ScheduledItem> scheduledItems = dataSnapshot.getValue(typeToken);
-                data.postValue(scheduledItems);
-            }
-
-            @Override
-            public void onCancelled(@NonNull DatabaseError databaseError) {
-                Log.w(TAG, "Error when getting schedule for event " + eventId, databaseError.toException());
-            }
-        });
-
-        return data;
+        return this.getElems(eventId);
     }
 }
