+++ app/src/main/java/ch/epfl/sweng/eventmanager/repository/SpotsRepository.java
@@ -1,12 +1,8 @@
 package ch.epfl.sweng.eventmanager.repository;
 
 import android.arch.lifecycle.LiveData;
-import android.arch.lifecycle.MutableLiveData;
-import android.support.annotation.NonNull;
-import android.util.Log;
-import ch.epfl.sweng.eventmanager.repository.data.ScheduledItem;
 import ch.epfl.sweng.eventmanager.repository.data.Spot;
-import com.google.firebase.database.*;
+import com.google.firebase.database.GenericTypeIndicator;
 
 import javax.inject.Inject;
 import javax.inject.Singleton;
@@ -16,35 +12,16 @@ import java.util.List;
  * @author Louis Vialar
  */
 @Singleton
-public class SpotsRepository {
+public class SpotsRepository extends AbstractEventDataRepository<Spot> {
     private static final String TAG = "SpotsRepository";
 
     @Inject
     public SpotsRepository() {
+        super("spots", new GenericTypeIndicator<List<Spot>>() {
+        });
     }
 
     public LiveData<List<Spot>> getSpots(int eventId) {
-        final MutableLiveData<List<Spot>> data = new MutableLiveData<>();
-        DatabaseReference dbRef = FirebaseDatabase.getInstance()
-                .getReference("spots")
-                .child("event_" + eventId);
-
-        dbRef.addValueEventListener(new ValueEventListener() {
-            @Override
-            public void onDataChange(@NonNull DataSnapshot dataSnapshot) {
-                GenericTypeIndicator<List<Spot>> typeToken = new GenericTypeIndicator<List<Spot>>() {
-                };
-
-                List<Spot> spots = dataSnapshot.getValue(typeToken);
-                data.postValue(spots);
-            }
-
-            @Override
-            public void onCancelled(@NonNull DatabaseError databaseError) {
-                Log.w(TAG, "Error when getting schedule for event " + eventId, databaseError.toException());
-            }
-        });
-
-        return data;
+        return this.getElems(eventId);
     }
 }
